Change Date and Time field

+++ Mobile/app/src/main/java/com/clearchain/app/domain/model/AdminStats.kt
@@ -1,23 +1,27 @@
 package com.clearchain.app.domain.model
 
 data class AdminStats(
+    // Organization stats
     val totalOrganizations: Int,
     val totalGroceries: Int,
     val totalNgos: Int,
     val verifiedOrganizations: Int,
     val unverifiedOrganizations: Int,
-    
+
+    // Listing stats
     val totalListings: Int,
     val activeListings: Int,
     val reservedListings: Int,
-    
+
+    // Pickup request stats
     val totalPickupRequests: Int,
     val pendingRequests: Int,
     val approvedRequests: Int,
-    val readyRequests: Int,        // ✅ ADD
-    val rejectedRequests: Int,     // ✅ ADD
+    val readyRequests: Int,
+    val rejectedRequests: Int,
     val completedRequests: Int,
-    val cancelledRequests: Int,    // ✅ ADD
-    
-    val totalFoodSaved: Int
+    val cancelledRequests: Int,
+
+    // Food saved
+    val totalFoodSaved: Double
 )
